refactor(websocket): extract ChatTurnHandler from the runner
The T4 step of docs/websocket-runner-refactor-plan.md: chat was the largest of
the six jobs UnifiedWebSocketRunner did, and every one of its private members
was reachable from the other five. ChatTurnHandler now owns the turn counter
and abort controller, the per-thread approval state, the capability run, and
the three shapes a chat_message can take — an agent turn, a workflow run, or a
media generation. It reaches the connection only through ClientSession; what
another region owns arrives in its deps.

The ToolBridge instances stay with the socket and are injected (plan D4): a
dropped connection must be able to cancel every pending waiter without asking
the chat handler anything. Chat's use of the connection's job map goes through
a ChatJobAccess dep rather than the map itself (plan D2), backed today by four
adapters in the runner's constructor and re-pointed at JobExecutionManager when
that extraction lands.

handleChatMessage moves whole. Splitting its phases is a second reading of
1,150 lines that this diff cannot also carry, so it stays a follow-up as the
plan allows. The runner keeps delegating members for the call sites outside
chat — the turn lifecycle, the request-seq accessor direct inference reads, and
the three internals the existing suites drive — each marked transitional and
removed in T7/T8. The runner drops from 8,829 lines to 5,515.
